Refactored the code design for data schemas in types/ ; edited conference ( added v2) and oveview prompt; added a loading icon on evaluation section on results screen
