+++ backend/local_cache/649eb84df5877fde/status.json
@@ -1,0 +1,1 @@
+{"job_id": "649eb84df5877fde", "transcript_name": "Moody's Corporation Presents at Oppenheimer 28th Annual Technology, Internet & Communications Conference, Aug-11-2025 10_45 AM – Moody's Corporation – BamSEC.json", "current_stage": "completed", "stages": {"validating": "completed", "q_a_summary": "completed", "overview_summary": "completed", "summary_evaluation": "completed"}, "percent_complete": 100, "updated_at": "2025-08-31T10:12:45.031533", "input": {"call_type": "conference", "summary_length": "long", "filename": "Moody's Corporation Presents at Oppenheimer 28th Annual Technology, Internet & Communications Conference, Aug-11-2025 10_45 AM – Moody's Corporation – BamSEC.pdf"}}
