+++ backend/local_cache/712fb987edb44884/overview_summary.json
@@ -1,0 +1,1 @@
+{"metadata": {"model": "gpt-5-mini", "prompt_version": "version_2", "max_output_tokens": 9000, "input_tokens": 4714, "output_tokens": 1450, "reasoning_tokens": null, "finish_reason": "completed", "remaining_tokens": 195069, "time": 25, "total_time": 48}, "data": {"title": "Atlassian Corporation Q4 2025 Earnings Call Aug 07, 2025", "executives_list": [{"executive_name": "Michael Cannon-Brookes", "role": "CEO"}, {"executive_name": "Joseph L. Binz", "role": "CFO"}, {"executive_name": "Martin Lam", "role": "Head IR"}], "overview": "Management highlighted strong FY '25 execution with over $5.2 billion in revenue and over $1.4 billion in free cash flow, continued Rule of 40 plus performance, and pronounced enterprise momentum evidenced by a record number of deals greater than $1 million in ACV in the quarter (up over 2x year-on-year) and a cloud net revenue retention rate of 120%. AI adoption was emphasized as a key growth driver—\"We reached 2.3 million AI now, up 50% from last quarter\"—and the newly launched Team '25/Teamwork Collection is exceeding expectations, driving consolidation onto Atlassian Cloud; management also announced a strategic Google Cloud partnership to accelerate cloud transformation and AI delivery. Leadership noted strong cross-product revenue growth (Jira, Confluence, Jira Service Management, Loom), large-customer migrations momentum, continued hiring and investment to pursue the company’s three transformations (Enterprise, AI, System of Work), and disclosed a planned President transition at year-end.", "guidance_outlook": null, "financial_results": null}}
