+++ backend/local_cache/712fb987edb44884/status.json
@@ -1,0 +1,1 @@
+{"job_id": "712fb987edb44884", "transcript_name": "Atlassian Corporation, Q4 2025 Earnings Call, Aug 07, 2025 – Atlassian Corporation – BamSEC.json", "current_stage": "completed", "stages": {"validating": "completed", "q_a_summary": "completed", "overview_summary": "completed", "summary_evaluation": "completed"}, "percent_complete": 100, "updated_at": "2025-08-31T10:29:06.905615", "input": {"call_type": "earnings", "summary_length": "short", "filename": "Atlassian Corporation, Q4 2025 Earnings Call, Aug 07, 2025 – Atlassian Corporation – BamSEC.pdf"}}
