--- backend/local_cache/a5e0773f4511524c/status.json
@@ -1,1 +1,0 @@
-{"job_id": "a5e0773f4511524c", "transcript_name": "Moody's Corporation Presents at Oppenheimer 28th Annual Technology, Internet & Communications Conference, Aug-11-2025 10_45 AM – Moody's Corporation – BamSEC.json", "current_stage": "completed", "stages": {"validating": "completed", "q_a_summary": "completed", "overview_summary": "completed", "summary_evaluation": "completed"}, "percent_complete": 100, "updated_at": "2025-08-30T13:53:27.513070", "input": {"call_type": "conference", "summary_length": "long", "filename": "Moody's Corporation Presents at Oppenheimer 28th Annual Technology, Internet & Communications Conference, Aug-11-2025 10_45 AM – Moody's Corporation – BamSEC.pdf"}}
