+++ backend/local_cache/a7abc824d8594c9c/status.json
@@ -1,0 +1,1 @@
+{"job_id": "a7abc824d8594c9c", "transcript_name": "Atlassian Corporation, Q4 2025 Earnings Call, Aug 07, 2025 – Atlassian Corporation – BamSEC.json", "current_stage": "failed", "stages": {"validating": "completed", "q_a_summary": "failed", "overview_summary": "pending", "summary_evaluation": "pending"}, "percent_complete": 25, "updated_at": "2025-08-31T10:21:47.388517", "input": {"call_type": "earnings", "summary_length": "short", "filename": "Atlassian Corporation, Q4 2025 Earnings Call, Aug 07, 2025 – Atlassian Corporation – BamSEC.pdf"}, "error": {"code": "llm_summary_error", "message": "OpenAI API error: Connection error."}, "warnings": ["Q&A summary failed: OpenAI API error: Connection error."]}
