--- backend/local_cache/f4e02997299d77a9/overview_summary.json
@@ -1,1 +1,0 @@
-{"metadata": {"model": "gpt-5-mini", "prompt_version": "version_2", "max_output_tokens": 5000, "input_tokens": 5521, "output_tokens": 1381, "reasoning_tokens": null, "finish_reason": "completed", "remaining_tokens": 194262, "time": 23, "total_time": 61}, "data": {"title": "Atlassian Q4 2025 Earnings Call Aug 07, 2025", "executives_list": [{"executive_name": "Joseph L. Binz", "role": "CFO"}, {"executive_name": "Martin Lam", "role": "Head IR"}, {"executive_name": "Michael Cannon-Brookes", "role": "CEO"}], "overview": "Atlassian closed FY'25 with $5.2 billion in revenue and $1.4 billion in free cash flow, highlighting strong enterprise execution—a record number of deals >$1 million in ACV (up >2x YoY) and cloud net revenue retention of 120%—while product momentum was broad-based with Jira, Confluence, Jira Service Management and Loom growing in line with or faster than the company. AI adoption accelerated (2.3 million AI users, up 50% QoQ; token usage up 5x QoQ) and Atlassian reported early Teamwork Collection (Team '25) traction exceeding expectations; management emphasized AI, enterprise expansion and the Atlassian system of work as the three strategic transformations, and announced a Google Cloud partnership to advance multi-cloud AI infrastructure. Leadership noted continued focus on cloud migrations and enterprise go-to-market as growth levers, and disclosed an executive transition with Anu departing as President at year end.", "guidance_outlook": null, "financial_results": null}}
