--- backend/local_cache/f4e02997299d77a9/status.json
@@ -1,1 +1,0 @@
-{"job_id": "f4e02997299d77a9", "transcript_name": "Atlassian Corporation, Q4 2025 Earnings Call, Aug 07, 2025 – Atlassian Corporation – BamSEC.json", "current_stage": "completed", "stages": {"validating": "completed", "q_a_summary": "completed", "overview_summary": "completed", "summary_evaluation": "completed"}, "percent_complete": 100, "updated_at": "2025-08-30T14:23:50.307678", "input": {"call_type": "earnings", "summary_length": "long", "filename": "Atlassian Corporation, Q4 2025 Earnings Call, Aug 07, 2025 – Atlassian Corporation – BamSEC.pdf"}}
