+++ backend/src/config/runtime.py
@@ -18,7 +18,7 @@ EFFORT_LEVEL_Q_A = "minimal"
 
 # CONFERENCE
 CONFERENCE_Q_A_MODEL = "gpt-5"
-CONFERENCE_LONG_Q_A_PROMPT_VERSION = "version_1"
+CONFERENCE_LONG_Q_A_PROMPT_VERSION = "version_2"
 
 EFFORT_LEVEL_Q_A_CONFERENCE = "minimal"
