+++ summarizer-app/src/services/api.ts
@@ -2,11 +2,11 @@ import axios from "axios"
 import { DocumentPickerAsset } from "expo-document-picker"
 import { API_BASE } from "../env" // keep as-is for relative path stability
 
-// !! IMPORTANTE !!
-
-// Substitua 'YOUR_COMPUTER_IP' pelo endereço IP da sua máquina na sua rede Wi-Fi.
-// Emuladores/dispositivos não conseguem acessar 'localhost' ou '127.0.0.1'.
-// Windows: `ipconfig` | macOS/Linux: `ifconfig`
+import {
+  ValidationResponse,
+  SummaryResponse,
+  CancelJobResponse,
+} from "../types/api"
 
 const apiClient = axios.create({
   baseURL: API_BASE,
@@ -41,52 +41,24 @@ export const validatePdf = async (
     // Send POST API request
     const response = await apiClient.post("/validate_file", formData)
     // Parse the response
-    return response.data as {
-      is_validated: boolean
-      validated_at?: string
-      input?: { call_type?: string; summary_length?: string; filename?: string }
-      transcript_name?: string
-      filename?: string
-      job_id: string
-    }
+    return response.data as ValidationResponse
   } catch (error) {
     throw error
   }
 }
 
-export const getSummary = async (jobId: string) => {
+export const getSummary = async (jobId: string): Promise<SummaryResponse> => {
   // GET request with a job id
   const response = await apiClient.get("/summary", {
     params: { job_id: jobId },
   })
 
-  return response.data as {
-    job_id: string
-    transcript_name: string
-    current_stage:
-      | "q_a_summary"
-      | "overview_summary"
-      | "summary_evaluation"
-      | "completed"
-      | "failed"
-    //each stage has these status: pending, running, completed, failed
-    stages: Record<string, "pending" | "running" | "completed" | "failed">
-    // for progress bar
-    percent_complete: number
-    updated_at: string
-    input?: { call_type?: string; summary_length?: string; filename?: string }
-    outputs?: {
-      q_a_summary?: any
-      overview_summary?: any
-      summary_evaluation?: any
-    }
-    error?: { code?: string; message?: string }
-  }
+  return response.data as SummaryResponse
 }
 
-export const cancelJob = async (jobId: string) => {
+export const cancelJob = async (jobId: string): Promise<CancelJobResponse> => {
   const response = await apiClient.post("/cancel", null, {
     params: { job_id: jobId },
   })
-  return response.data as { ok: boolean; job_id: string; status: string }
+  return response.data as CancelJobResponse
 }
